Remove deploy_solidity_contract and dependencies
